==> 0.merged.R <==
library(tidyverse)
library(openxlsx)
library(readxl)
library(dplyr)
library(lubridate)

####1.check####
fcg.check <- read_excel("fcg.xlsx", sheet = "check")
qz.check <- read_excel("qinzhou.xlsx", sheet = "check")
cz.check <- read_excel("chongzuo.xlsx", sheet = "check")
nn.check <- read_excel("nn.xlsx", sheet = "check")
str(fcg.check)
str(qz.check)
str(cz.check)
str(nn.check)
check <- bind_rows(fcg.check, qz.check, cz.check, nn.check)
check$ID <- gsub("'", '', check$ID)
check <- check %>% select(1,2,6,7,everything())
write.xlsx(check, "check.xlsx")

####2.info####
fcg.art.info <- read_excel("fcg.xlsx", sheet = "art.info")
qz.art.info <- read_excel("qinzhou.xlsx", sheet = "art.info")
cz.art.info <- read_excel("chongzuo.xlsx", sheet = "art.info")
nn.art.info <- read_excel("nn.xlsx", sheet = "art.info")
art.info <- bind_rows(fcg.art.info, qz.art.info, cz.art.info, nn.art.info)  
art.info$ID <- gsub("'", '',art.info$ID)
art.info <- art.info %>% group_by(ID) %>% distinct(ID, .keep_all = TRUE) %>% ungroup()

fcg.flw.info <- read_excel("fcg.xlsx", sheet = "flw.info") %>% ungroup()
qz.flw.info <- read_excel("qinzhou.xlsx", sheet = "flw.info") %>% ungroup()
cz.flw.info <- read_excel("chongzuo.xlsx", sheet = "flw.info")%>% ungroup()
nn.flw.info <- read_excel("nn.xlsx", sheet = "flw.info") %>% ungroup()
nn.flw.info$ID <- gsub("A", '', nn.flw.info$ID )
str(fcg.flw.info)
str(qz.flw.info)
str(cz.flw.info)
str(nn.flw.info)
qz.flw.info$DiagnosedDate <- as.Date(qz.flw.info$DiagnosedDate)
cz.flw.info$DiagnosedDate <- as.Date(cz.flw.info$DiagnosedDate)
flw.info <- bind_rows(fcg.flw.info, qz.flw.info, cz.flw.info, nn.flw.info)
flw.info$ID <- gsub("'", '', flw.info$ID)
flw.info <- flw.info %>% group_by(ID) %>% distinct(ID, .keep_all = TRUE) %>% ungroup()

info <- art.info %>% left_join(flw.info, by = "ID")
info <- info %>% select(1,2,4,8,19, everything())
info <- info %>% select(1:9,11,17:19, everything())
write.xlsx(info, "info.xlsx")

####3.check.1####
info.select <- info %>% select(1,3,4)
str(info.select)
info.select <- info.select %>% mutate_at(vars(ends_with("Date")),as.Date)
info.select <- info.select %>%
  mutate(ARTinitialAge = floor(as.numeric(difftime(ARTinitialDate, BirthDate, units = "days"))/365.25))

check <- check %>% left_join(info.select, by = "ID")
check <- check %>% mutate_at(vars(ends_with("Date")), as.Date)
str(check)
check <- check %>% filter(!is.na(ARTinitialDate))
check <- check %>% filter(ARTinitialAge>=18)
check <- check %>% select(1,11,12,13, everything())

check$VL <- as.numeric(check$VL)
check$Gly <- as.numeric(check$Gly)
str(check)
write.xlsx(check, "0.check.xlsx")


fcg.death <- read_excel("fcg.xlsx", sheet = "death")
qz.death <- read_excel("qinzhou.xlsx", sheet = "death")
cz.death <- read_excel("chongzuo.xlsx", sheet = "death")
nn.death <- read_excel("nn.xlsx", sheet = "death")
str(fcg.death)
str(qz.death)
str(cz.death)
str(nn.death)
death <- bind_rows(fcg.death, qz.death, cz.death, nn.death)
death <- death %>% group_by(ID) %>% distinct(ID, .keep_all = TRUE) %>% ungroup()
death$ID <- gsub("'", '', death$ID)
death$DiedDate <- as.Date(death$DiedDate)
str(death)
write.xlsx(death, "death.xlsx")
